change char's xadvance property from uint16 to float
(cherry picked from commit ebd50a106b43b3f490e60fc74966f7eebfa155fe)
(cherry picked from commit ef5452ff7c2e601f5c27185c27f3b74ba4880be3)
(cherry picked from commit f576284c908f9761947db7a6c64201a4e50a8a12)

+++ Source/LGUI/Private/Core/UIGeometry.cpp
@@ -731,7 +731,7 @@ void UIGeometry::UpdateUIText(const FString& text, int32 visibleCharCount, float
 		else
 		{
 			auto charData = font->GetCharData(charCode, overrideFontSize);
-			return (float)charData->xadvance;
+			return charData->xadvance;
 		}
 	};
 	auto GetUnderlineOrStrikethroughCharGeo = [&](TCHAR charCode, int overrideFontSize)
